XML Saving added, Button depressed option
Implementied saving to XML, not fully wired in yet with a button. On
that note, expanded on the button class to allow a depressed image
rather than just a scale in sort of deal.

XML script loading not implemented, and the save script will have to
change to accomodate conditionals.